Separate forex and equity endpoints for tradermade adapter
