+++ packages/sources/tradermade/src/adapter.ts
@@ -4,37 +4,18 @@ import {
   MakeWSHandler,
   AdapterRequest,
   ExecuteFactory,
+  APIEndpoint,
 } from '@chainlink/types'
-import { Requester, Validator } from '@chainlink/ea-bootstrap'
-import { makeConfig, DEFAULT_WS_API_ENDPOINT, NAME } from './config'
+import { Requester, Validator, Builder } from '@chainlink/ea-bootstrap'
+import { makeConfig, DEFAULT_WS_API_ENDPOINT } from './config'
+import * as endpoints from './endpoint'
 
-const customParams = {
-  base: ['base', 'from', 'symbol', 'market'],
-  quote: ['quote', 'to', 'market', 'convert'],
+export const execute: ExecuteWithConfig<Config> = async (request, context, config) => {
+  return Builder.buildSelector(request, context, config, endpoints)
 }
 
-export const execute: ExecuteWithConfig<Config> = async (input, _, config) => {
-  const validator = new Validator(input, customParams)
-  if (validator.error) throw validator.error
-
-  Requester.logConfig(config)
-
-  const jobRunID = validator.validated.id
-  const symbol = (validator.overrideSymbol(NAME) as string).toUpperCase()
-  const quote = (validator.validated.data.quote || '').toUpperCase()
-  const currency = `${symbol}${quote}`
-
-  const params = {
-    ...config.api.params,
-    currency,
-  }
-
-  const options = { ...config.api, params }
-
-  const response = await Requester.request(options)
-  response.data.result = Requester.validateResultNumber(response.data, ['quotes', 0, 'mid'])
-  return Requester.success(jobRunID, response)
-}
+export const endpointSelector = (request: AdapterRequest): APIEndpoint =>
+  Builder.selectEndpoint(request, makeConfig(), endpoints)
 
 export const makeExecute: ExecuteFactory<Config> = (config) => {
   return async (request, context) => execute(request, context, config || makeConfig())
@@ -51,7 +32,7 @@ export const makeWSHandler = (config?: Config): MakeWSHandler => {
     return sub
   }
   const getPair = (input: AdapterRequest) => {
-    const validator = new Validator(input, customParams, {}, false)
+    const validator = new Validator(input, endpoints.forex.customParams, {}, false)
     if (validator.error) return
     const base = validator.validated.data.base.toUpperCase()
     const quote = validator.validated.data.quote.toUpperCase()
@@ -63,6 +44,8 @@ export const makeWSHandler = (config?: Config): MakeWSHandler => {
       connection: {
         url: defaultConfig.api.baseWsURL || DEFAULT_WS_API_ENDPOINT,
       },
+      shouldNotServeInputUsingWS: (input: AdapterRequest) =>
+        endpoints.forex.supportedEndpoints.indexOf(input.data.endpoint) === -1,
       subscribe: (input: AdapterRequest) => getSubscription(getPair(input)),
       unsubscribe: () => null, // Tradermade does not support unsubscribing.
       subsFromMessage: (message) => {
