+++ packages/sources/tradermade/src/endpoint/index.ts
@@ -1,0 +1,2 @@
+export * as live from './live'
+export * as forex from './forex'
